ttest_xblock/requirement.txt
absl-py==1.2.0
acid-xblock==0.2.1
appdirs==1.4.4
arrow==1.2.2
asgiref==3.5.2
astroid==2.11.6
attrs==20.3.0
bcrypt==3.1.7
binaryornot==0.4.4
bok-choy==0.7.1
boto==2.49.0
boto3==1.24.10
botocore==1.27.10
cached-property==1.5.2
certifi==2022.6.15
cffi==1.14.3
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
click-log==0.4.0
code-annotations==1.3.0
codecov==2.1.12
cookiecutter==2.1.1
coverage==6.4.1
cryptography==3.2.1
cycler==0.11.0
ddt==1.5.0
diff-cover==4.0.0
dill==0.3.5.1
distlib==0.3.4
distro==1.5.0
Django==3.2.13
django-pyfs==3.2.0
# Editable install with no version control (django-stream-xblock==0.1)
docker==4.3.1
docker-compose==1.27.4
dockerpty==0.4.1
docopt==0.6.2
edx-lint==5.2.4
filelock==3.7.1
Flask==2.1.3
Flask-Uploads==0.2.1
fonttools==4.34.4
fs==2.4.16
fs-s3fs==1.1.1
hypothesis==6.47.0
idna==2.10
importlib-metadata==4.3.0
imutils==0.5.4
inflect==5.6.0
iniconfig==1.1.1
isort==5.10.1
Jinja2==3.1.2
jinja2-pluralize==0.3.0
jinja2-time==0.2.0
jmespath==1.0.0
joblib==1.1.0
jsonschema==3.2.0
kiwisolver==1.4.4
lazy==1.4
lazy-object-proxy==1.7.1
lxml==4.9.0
Mako==1.2.0
matplotlib==3.5.2
mccabe==0.7.0
mediapipe==0.8.10.1
mock==4.0.3
needle==0.5.0
nose==1.3.7
opencv-contrib-python==4.6.0.66
opencv-python==4.5.4.60
pandas==1.4.3
paramiko==2.7.2
path==16.4.0
pbr==5.9.0
pep517==0.12.0
Pillow==9.1.1
pip-tools==6.6.2
platformdirs==2.5.2
pluggy==1.0.0
protobuf==3.20.1
py==1.11.0
pycodestyle==2.8.0
pycparser==2.20
pydocstyle==6.1.1
Pygments==2.12.0
pylint==2.14.2
pylint-celery==0.3
pylint-django==2.5.3
pylint-plugin-utils==0.7
pymongo==4.2.0
PyNaCl==1.4.0
pyparsing==3.0.9
pypng==0.0.21
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyQt6==6.3.1
PyQt6-Qt6==6.3.1
PyQt6-sip==13.4.0
pyrsistent==0.17.3
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-rerunfailures==10.2
python-dotenv==0.15.0
python-slugify==6.1.2
PyYAML==5.3.1
requests==2.24.0
s3transfer==0.6.0
scikit-learn==1.1.2
seaborn==0.11.2
selenium==3.4.1
simplejson==3.17.6
six==1.15.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
stevedore==3.5.0
text-unidecode==1.3
texttable==1.6.3
thop==0.1.1.post2207130030
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
torchsummary==1.5.1
tox==3.25.0
tox-battery==0.6.1
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.25.11
virtualenv==20.14.1
web-fragments==2.0.0
WebOb==1.8.7
websocket-client==0.57.0
Werkzeug==0.16.0
wrapt==1.14.1
zipp==1.2.0
